Фиксы

+++ app/User.php
@@ -54,6 +54,9 @@ use Storage;
  * @method static \Illuminate\Database\Eloquent\Builder|\App\User whereUpdatedAt($value)
  * @property-read \Illuminate\Database\Eloquent\Collection|\App\Review[] $reviews
  * @property-read \Illuminate\Database\Eloquent\Collection|\App\ReviewEstimate[] $reviewEstimates
+ * @method static \Illuminate\Database\Eloquent\Builder|\App\User whereAbout($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\App\User whereApiToken($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\App\User whereTimezone($value)
  */
 class User extends Authenticatable
 {
@@ -73,9 +76,10 @@ class User extends Authenticatable
      * @var array
      */
     protected $fillable = [
-        'nickname', 'email', 'surname',
-        'patronymic', 'birthday_date', 'gender',
-        'about', 'timezone',
+        'nickname', 'email', 'name',
+        'surname', 'patronymic', 'birthday_date',
+        'gender', 'about', 'timezone',
+        'password', 'avatar',
     ];
 
     /**
@@ -147,31 +151,22 @@ class User extends Authenticatable
      * Установить аватар для пользователя
      *
      * @param UploadedFile $avatar Аватар пользователя
-     * @param bool $save Сохранять ли состояние модели после записи
      * @return void
      * @throws Exception
      */
-    public function setAvatar(UploadedFile $avatar, bool $save = false)
+    public function setAvatarAttribute(UploadedFile $avatar)
     {
-        if (blank($this->id) && !$save) {
+        if (blank($this->id)) {
             throw new Exception('For setting avatar path, user must be present');
         }
 
-        if ($save) {
-            $this->save();
-        }
-
-        if (Storage::disk('s3')->exists($this->avatar_path)) {
+        if (filled($this->avatar) && Storage::disk('s3')->exists($this->avatar_path)) {
             Storage::disk('s3')->delete($this->avatar_path);
         }
 
         $image_name = $this::AVATAR_PATH . '/' . $this->id;
         $storage_path = Storage::disk('s3')->put($image_name, $avatar);
-        $this->avatar = basename($storage_path);
-
-        if ($save) {
-            $this->save();
-        }
+        $this->attributes['avatar'] = basename($storage_path);
     }
 
     /**
@@ -187,15 +182,17 @@ class User extends Authenticatable
 
         switch ($this->gender) {
             case $this::GENDER_MALE:
-                return '/img/avatar_man.png';
+                $avatar_url = '/img/avatar_man.png';
                 break;
             case $this::GENDER_FEMALE:
-                return '/img/avatar_woman.png';
+                $avatar_url = '/img/avatar_woman.png';
                 break;
             default:
-                return '/img/avatar_default.png';
+                $avatar_url = '/img/avatar_default.png';
                 break;
         }
+
+        return $avatar_url;
     }
 
     /**
@@ -340,5 +337,15 @@ class User extends Authenticatable
     public function getBirthdayDatePlainAttribute()
     {
         return $this->attributes['birthday_date'];
+    }
+
+    /**
+     * Записать хэш нового пароля пользователя
+     *
+     * @param string $password
+     */
+    public function setPasswordAttribute($password)
+    {
+        $this->attributes['password'] = bcrypt($password);
     }
 }
